feat(redis): Parallel sync and permission repository
The sync process is currently single threaded which is a bottle neck
for large data sets. This change adds in a worker pool and some
parallel streaming.

The worker pool size can be configured with

```yaml
fiat:
  worker:
    parallelism: 4
```

Care must be taken to ensure the Redis connection pool size is large
enough to cope with the sync process and additional real time requests.